Work around cellpose bug and improve console logging
Cellpose runs "squeeze" on a certain array during an internal step in the
algorithm to eliminate some unneeded singleton dimensions, but when the input
image has a legitimate 1-sized dimension this is inadvertently squeezed out as
well triggering an exception when the array dimensionality is not what's
expected. We can work around this by trimming the image dimensions slightly
when computing our rolling window coordinates without any impact on our
results.

In addition to working around that issue, we add some extra logging around
the window overlap size and auto-contrast computation.